+++ src/DependencyInjection/GoogleMapsExtension.php
@@ -28,7 +28,6 @@ class GoogleMapsExtension extends Extension
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
 
         $loader->load('commands.yml');
-        $loader->load('listeners.yml');
         $loader->load('services.yml');
         $loader->load('twig.yml');
     }